changed interface for segment tree
The reason for this is that the updates made use of __setitem__, making
the update procedure opaque. This uses a dedicated 'set_values' method
to make it clear to the user that this is the entry point for updating
the segment tree.

The __getitem__ was also updated. It used to provide access to
tree.values. The new interface provides a lookup like

  tree[level, index]

where 'level' is the level in the tree and 'index' is the value index
within that level.